requirements.txt
Django==2.2.10
djangorestframework==3.12.4
python-status==1.0.1
pytz==2021.1
sqlparse==0.4.1
